Add a *dynamic* patch for all code that throws SSL verification errors
This is crazy, but it seems to work! The case in point example here is
Vimeo, which is heavily obfuscated (so we can't reliably match method or
class names) but uses OkHttp internally, which uses the standard
exception types. We can spot those, so we do retrospective patching: the
first time a certificate validation fails, we disable the method that
threw the exception.

We'll still get one initial failure, but after that everything works
nicely. Wild.