Disable unnecessary AbstractVerifier patch
This only applied to a specific backport of Apache's HttpClient to
Android, from 2014, so this is pretty niche anyway.

Regardless of that, this doesn't actually do cert pinning - it just
disables all normal checks entirely. These scripts are intended to be
used for a working interception setup, so those shouldn't be required.
AbstractVerifier (in every version I can find) doesn't support check for
a specific cert or CA at all.