Revert "Always loudly log intercepted connections that fail to reach the proxy"
This reverts commit 6eec741f5c6842916a9158c2d9c98592703dfb16.

This doensn't actually work correctly, as connect() will return -1
(treated here as failure) for sockets that are still in progress, when
opened non-blocking, and therefore we end up logging these as failures.
We need to handle async connection state detection - that's a bit fiddly
from inside Frida, so for now let's just roll this back.